Use a single assertion
The assertions were failing, but most of the assertions were failing on
the code being incorrect. Since STDOUT and STDERR tend to provide more
clues as to what it going wrong, we move all the assertions to be a
single assertion so we can see the output of STDOUT and STDERR when
there's a failure.